requirements.txt
apiclient==1.0.4
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
Django==4.2.11
django-rest-framework==0.1.0
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
exceptiongroup==1.2.1
gunicorn==22.0.0
idna==3.6
iniconfig==2.0.0
lxml==5.2.1
marshmallow==3.17.0
mysqlclient==2.2.4
packaging==24.0
pillow==10.3.0
pluggy==1.5.0
PyJWT==2.8.0
pytest==8.2.2
pytest-django==4.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.31.0
spyne==2.14.0
sqlparse==0.4.4
suds==1.1.2
tomli==2.0.1
transbank-sdk==5.0.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.7.0
